Refactor hn and memory jogger code to separate modules
